Rename export/index.js to export-notes.js
