+++ lib/utils/export/to-zip.js
@@ -1,7 +1,7 @@
 import sanitize from 'sanitize-filename';
 import { identity, update } from 'lodash';
 
-import { LF_ONLY_NEWLINES } from './';
+import { LF_ONLY_NEWLINES } from './export-notes';
 
 const FILENAME_LENGTH = 40;
 const TAG_LINE_LENGTH = 75;
